tests: Move `GET /crates/:name/downloads` tests to `krate::downloads` module
